tirando o int e colocando o float

+++ UsandoRepetição.py
@@ -1,7 +1,7 @@
 
 
 print('Digite a resposta da soma de 20 + 30')
-r = int(input('Resposta:'))
+r = float(input('Resposta:'))
 
 if r == 50:
         print('Resposta CORRETA!')
@@ -12,4 +12,3 @@ elif r != 50:
 print()
 print('FIM DO PROGRAMA!')
 print()
-
